.NET Framework 4.8 Implementation
