+++ IT Returns Label Maker/IT Returns Label Maker.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32929.385
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IT Returns Label Maker", "IT Returns Label Maker\IT Returns Label Maker.csproj", "{098F50A4-FB44-4FC3-87B2-30101F24F005}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{098F50A4-FB44-4FC3-87B2-30101F24F005}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{098F50A4-FB44-4FC3-87B2-30101F24F005}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{098F50A4-FB44-4FC3-87B2-30101F24F005}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{098F50A4-FB44-4FC3-87B2-30101F24F005}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6F878E6A-6233-4861-A5E9-6A5A3A313EA1}
+	EndGlobalSection
+EndGlobal
