+++ IT Returns Label Maker/IT Returns Label Maker/Program.cs
@@ -1,23 +1,22 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace ReturnLabelMaker
+namespace IT_Returns_Label_Maker
 {
-    static class Program
+    internal static class Program
     {
         /// <summary>
-        ///  The main entry point for the application.
+        /// The main entry point for the application.
         /// </summary>
         [STAThread]
         static void Main()
         {
-            Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new ReturnsForm());
+            Application.Run(new Form1());
         }
     }
 }
