+++ IT Returns Label Maker/IT Returns Label Maker/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net48" />
+  <package id="RestSharp" version="106.15.0" targetFramework="net48" />
+</packages>
